+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Next Release
 
+#### Improvements
+
+- Add `AUDITLOG_USE_BASE_MANAGER` setting to override default manager use ([#766](https://github.com/jazzband/django-auditlog/pull/766))
+- Drop 'Python 3.9' support ([#773](https://github.com/jazzband/django-auditlog/pull/773))
+
 ## 3.3.0 (2025-09-18)
 
 #### Improvements